Newest version
In the crootgenerate tab my vlookup is failing. Even writing the most simple block (shown in highlighted cells) is failing and I can't figure out why. Taking a break for a while and if I start working on it again today I'm just going to replace the vlookup with a VBA code snippet